[RAMEN9610-12171] ion: fix stack-out-of-bounds
The align variable is 4bytes but casted with 8bytes,
so it makes stack-out-of-bounds problem.

+++ drivers/staging/android/ion/ion_buffer_protect.c
@@ -39,7 +39,7 @@ static unsigned long find_first_fit_with_align(unsigned long *map,
 					       unsigned int nr, void *data,
 					       struct gen_pool *pool)
 {
-	unsigned long align = ((*(unsigned long *)data) >> PAGE_SHIFT);
+	unsigned long align = ((*(unsigned int *)data) >> PAGE_SHIFT);
 
 	if (align > (1 << MAX_IOVA_ALIGNMENT))
 		align = (1 << MAX_IOVA_ALIGNMENT);
